Refine some of the output from mailer.py
* hook-scripts/mailer.py
  (StandardOutput.run_diff): ensure that we flush stdout/stderr before
    forking so the parent/child stay in sync.
  (_select_deletes, _select_modifies): test against None rather than
    "not true". the root directory is the empty string, which is
    considered false.
  (generate_list): report property (and text) changes, along with
    copies in the summary of the message. makes use of the new
    .text_changed property on the _change objects.
  (generate_diff): stop printing directory copies; that is now in the
    summary. make sure newlines are printed before each diff's header.
    don't print Copied lines for unchanged files (that info is in the
    summary). don't print diffs if the text didn't change (i.e. the
    props were changed, but no text change). use the new
    .either_binary() method on the FileDiff to avoid trying to diff
    binary file(s); print an appropriate message.
  (class _change): add a new 'text_changed' attribute. changed all
    constructor calls as appropriate.
  (ChangeCollector.apply_textdelta): if a change already exists, then
    ensure that its text_changed attribute gets set.

* hook-scripts/tests/mailer-t1.sh: add a usage message when missing
    args. print a success message for clarity.

* hook-scripts/tests/mailer-t1.output: update the reference output.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@844419 13f79535-47bb-0310-9956-ffa450edef68

+++ tools/hook-scripts/tests/mailer-t1.sh
@@ -12,6 +12,11 @@
 # USAGE: ./mailer-t1.sh REPOS MAILER-SCRIPT
 #
 
+if test "$#" != 2; then
+    echo "USAGE: ./mailer-t1.sh REPOS MAILER-SCRIPT"
+    exit 1
+fi
+
 glom=mailer-t1.current
 rm -f $glom
 
@@ -23,4 +28,4 @@ done
 echo "current mailer.py output in: $glom"
 
 echo diff -q mailer-t1.output $glom
-diff -q mailer-t1.output $glom
+diff -q mailer-t1.output $glom && echo "SUCCESS: no differences detected"
